Fix TypeError for PyPy
A module object is not subscriptable in PyPy. A subscriptable module is an implementation detail for CPython. The help doc only says you can use dotted notation to get attributes out of a module, it doesn't implement __getitem__. One should also not touch __builtins__ as it is an implementation detail for CPython.