Change KeyError into AttributeError in the comparator_factory of CompositeType
In order to support the 'generative' style of select() of sqlalchemy>=2, the
select() function in sqlalchemy==1.4 checks whether the first argument is an
iterable or not. This check is performed using 'hasattr', which actually calls
'getattr' and catch the AttributeError exceptions. Changing the exception
raised in the comparator_factory thus allows the 'select()' function to detect
that the CompositeType should be processed as a clause, not as a list of
clauses.